Separated jms into 2 modules: for jms-1 and jms-2. Otherwise instrumentation doesn't work.
