+++ plugins/opentracing-specialagent-jms2/src/main/java/io/opentracing/contrib/specialagent/jms2/Jms2AgentIntercept.java
@@ -12,15 +12,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.specialagent.jms;
+package io.opentracing.contrib.specialagent.jms2;
 
+import io.opentracing.contrib.jms.common.TracingMessageConsumer;
+import io.opentracing.contrib.jms2.TracingMessageProducer;
 import javax.jms.MessageConsumer;
 import javax.jms.MessageProducer;
 
-import io.opentracing.contrib.jms.TracingMessageProducer;
-import io.opentracing.contrib.jms.common.TracingMessageConsumer;
-
-public class JmsAgentIntercept {
+public class Jms2AgentIntercept {
   public static Object createProducer(final Object thiz) {
     return thiz instanceof TracingMessageProducer ? thiz : new TracingMessageProducer((MessageProducer)thiz);
   }
