+++ plugins/opentracing-specialagent-jms2/src/test/java/io/opentracing/contrib/specialagent/jms2/JmsTest.java
@@ -12,17 +12,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.specialagent.jms;
+package io.opentracing.contrib.specialagent.jms2;
 
-import static org.awaitility.Awaitility.*;
-import static org.hamcrest.core.IsEqual.*;
-import static org.junit.Assert.*;
+import static org.awaitility.Awaitility.await;
+import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
+import io.opentracing.contrib.jms.common.TracingMessageConsumer;
+import io.opentracing.contrib.jms2.TracingMessageProducer;
+import io.opentracing.mock.MockSpan;
+import io.opentracing.mock.MockTracer;
 import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
-
 import javax.jms.Connection;
 import javax.jms.DeliveryMode;
 import javax.jms.Destination;
@@ -32,13 +36,7 @@ import javax.jms.MessageListener;
 import javax.jms.MessageProducer;
 import javax.jms.Session;
 import javax.jms.TextMessage;
-
 import org.junit.Test;
-
-import io.opentracing.contrib.jms.common.TracingMessageConsumer;
-import io.opentracing.contrib.jms.TracingMessageProducer;
-import io.opentracing.mock.MockSpan;
-import io.opentracing.mock.MockTracer;
 
 public abstract class JmsTest {
   static final Logger logger = Logger.getLogger(JmsTest.class.getName());
