+++ plugins/pom.xml
@@ -35,6 +35,7 @@
     <module>opentracing-specialagent-aws-sdk</module>
     <module>opentracing-specialagent-cassandra-driver</module>
     <module>opentracing-specialagent-jms</module>
+    <module>opentracing-specialagent-jms2</module>
   </modules>
   <profiles>
     <profile>
